requirements.txt
numpy==1.21.1
openpyxl==3.0.7
pandas==1.3.1
matplotlib==3.4.2
scikit-learn==0.24.2
joblib==1.0.1
PyYAML==5.4.1
